+++ src/interfaces/Dropdown.ts
@@ -2,6 +2,10 @@ export interface ItemRegistration {
   position: number;
 }
 
+export interface ItemKeyboardEvent {
+  keyboardEvent: KeyboardEvent;
+}
+
 export interface GroupRegistration {
   items: HTMLCalciteDropdownItemElement[];
   position: number;